fix/header
header widget destructuring
