+++ src/widgets/header/ui/Header.module.scss
@@ -9,18 +9,6 @@
    padding: 40px 0;
 }
 
-.right {
-   display: flex;
-   align-items: center;
-   gap: 50px;
-}
 
-nav > ul {
-   display: flex;
-   gap: 20px;
-}
 
-.buttons {
-   display: flex;
-   gap: 30px;
-}
+
