+++ src/widgets/header/ui/nav/Nav.module.scss
@@ -1,0 +1,4 @@
+.list {
+   display: flex;
+   gap: 20px;
+}
