+++ src/widgets/header/ui/rightSide/RightSide.module.scss
@@ -1,0 +1,5 @@
+.root {
+   display: flex;
+   align-items: center;
+   gap: 50px;
+}
