+++ src/widgets/header/ui/userLinks/UserLinks..module.scss
@@ -1,0 +1,4 @@
+.buttons {
+   display: flex;
+   gap: 30px;
+}
